prototype/apps/backend/onboarding/src/logic/handle_github/requirements.txt
functions-framework
firebase-admin
google-cloud-firestore
google-cloud-aiplatform
vertexai
requests